Reorganize the panel routes
