+++ client/src/panels/panel_routes/IsolatedPanel.yaml
