+++ client/src/panels/panel_routes/PanelInventory.js
