+++ client/src/panels/panel_routes/PanelInventory.yaml
